prov/sm2: Add HMEM Inject Support
This patch adds support for HMEM, and allows EFA to pick up SM2.

EFA is not picking up SM2 b/c it does not support FI_HMEM.
OMPI attempts to create a libfabric provider with HMEM support,
and if it fails, it tries to create one without HMEM support.
EFA is choosing to not hard-fail if SHM provider cannot be picked
up, so when we specify SM2 for EFA, we run without SHM.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>