prov/sm2: Add FI_DELIVERY_COMPLETE
This change adds FI_DELIVERY_COMPLETE support by including the context
in the xfer_entry. When we receive a returned xfer_entry, if
FI_DELIVERY_COMPLETE is in the ops_flags, we write a completion. This
patch cleans up how we handle op_flags, and we are able to remove some
old #defines.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>