prov/cxi: Handle SW matched and mrecv edge case
Before an mrecv buffer is appened to the priority list, it is matched
against the onloaded SW unexpected list. Each enter of the SW unexpected
list contains a PUT_EVENT.

If an mrecv matches a SW unexpected list entry, the single PUT_EVENT is
used to progress the mrecv operation twice. Thus, the event chain looks
like PUT_EVENT -> PUT_EVENT -> REPLY_EVENT. Back-to-back PUT_EVENTs will
cause rdzv_mrecv_req_lookup() to return -FI_EAGAIN will causes the app
to abort (with debug enabled).

For this case only, in rdzv_mrecv_req_lookup(), it is safe to ignore the
back-to-back put event checking. It is NOT safe to ignore back-to-back
event checking for hardware progressed mrecv.

NETCASSINI-5147

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>