contrib/intel/jenkins: Add slurm support to Jenkinsfile
Add Slurm support to jenkinsfile with groovy functions
We no longer use the node step in the Jenkinsfile because it got replaced with slurm.
Removed the "withEnv -> sh " steps in every stage and simplified with either a call to run_fabtest or run_middleware
Every stage should be able to use one of these generified functions with the exception of ze-shm stage.
Fields were added to the top of the Jenkinsfile to allow for easier changes to MPI usage and python version upgrades.
DAOS is disabled and will be added back after the migration

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>