prov/sm2: Align sm2 region start to page size
Remove the alignment of the list of allocation entries.  These are not
page-aligned entries anyways, and are not in the critical path.

Correct the alignment of the regions which hold the xfer entries.  The
alignment was previously based on universe size, but I believe that was
the result of an erroneous refactor.

Signed-off-by: Luke Robison <lrbison@amazon.com>