prov/efa: extend rxr_pkt_entry_send() to send multiple packets
This patch extended the function rxr_pkt_entry_send() such
that it can be used to send multiple packets.

It renamed the argument "pkt_entry" to "pkt_entry_vec", and
made it an array of packet entries.

It added an argument "pkt_entry_cnt" for number of packet
entries to send.

For now, all the callers call rxr_pkt_entry_send() using 1
packet entry.

Signed-off-by: Wei Zhang <wzam@amazon.com>