prov/tcp: Removed unused saved_msg::ep field to fix assert
An MPICH spawn test triggered the following assert:

    xnet_progress.c, line 100
    assert(!ep->saved_msg->ep)

Although it is unlikely the test is hitting this condition (meaning
the assert is coming from some other cause), it is possible for
a socket to close (error or peer exited), and a new connection to
take its place.  If this happens, the new ep would take this path
in xnet_save_and_cont() and trigger the assert.

It turns out that the saved_msg::ep field isn't used anywhere.
The value is only set.  So, we can remove the field and this
assertion.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>