[Intel/CI]: Removing duplicate env settings and clean up log directories.
Since sourcing of env variables is prepended to the launch command,
set_environment has become redundant and is removed. This also obfuscates
the need to set the env variables in the jenkinsfile and this is removed as well.

Added removing daos_logs before running the daos tests.This doesnot remove
the test logs.

Signed-off-by: Nikhil Nanal <nikhil.nanal@intel.com>