prov/tcp: Handle prefetched data after reporting ETRUNC error
When progressing receive data, we need to process all data that's
been read into the socket prefetch buffer.  Prefetched data won't
generate POLLIN events on the socket, since it's already been read
from the kernel.

In the case of handling a truncated message, we write an error to
the CQ, update the current rx msg to read off the rest of the
message data, then return EAGAIN back to the progress loop.  This
results in the progress loop exiting.  However, if there's data
in the prefetch buffer, we won't return to claim it until the
socket signals the POLLIN event.  In some MPICH tests, that doesn't
occur, resulting in the test hanging.

Instead of returning to the progress loop, continue trying to read
the message data and only return once the message completes or there's
no data left to read.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>