prov/sm2: Get rid of xfer_ctx_pool in unexpected recv path
Remove a legacy SHM construct, xfer_ctx_pool. For the unexpected path,
SHM needs to copy the message out of the circular buffer and into a
temporary context until a receive is posted.  SM2 doesn't need to do
this behavior.  SM2 pops the message from the fifo queue, and then can
hold onto the peer's xfer_entry until the receiving endpoint posts a
buffer that matches it. This change reduces the amount of code on SM2's
critical path, and speeds it up a little bit.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>