prov/shm: add memory barrier before update resp->status for SAR
This patch added an write memory barrier to function
smr_copy_to_sar() and smr_copy_from_sar before updating
resp->status.

This is because resp->status is used by other functions
as indicator for completion of copy. Without the memory barrier,
the copy might not complete when resp-status is updated, which
will lead to data corruption.

Signed-off-by: Wei Zhang <wzam@amazon.com>