prov/shm: lock entire progress loop
Previously, a lot of lock acquiring and releasing of locks was
happening within the EP progress loop. Having constant lock
management is confusing, prone to deadlocks, and much of it is
redundant for non-threaded applications. Since the peer srx
implementation requires locking around all progress anyway, moving
to a single acquire/release around the progress is better anyway.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>