fabtests: make tx_buf and rx_buf aligned to 64 bytes by default
Prior to this change, tx_buf and rx_buf is not aligned by default.
Moreover, tx_buf is usually on the 2nd half of the whole buf.
Therefore, any change to buf_size would change the alignment
of tx_buf, which leads to the performance change.

For example, recently fabtests added the data verificiaton of
rma_bw, for which 4 bytes was added to buf size, which shifted
tx_buf's alignment and changed performance.

To address this issue, this patch make tx_buf and rx_buf to
align on 64 bytes boundary by default.

Signed-off-by: Wei Zhang <wzam@amazon.com>