prov/tcp: Continue receiving after truncation error
If we receive a message and the matching buffer is too small,
rather than closing the connection, report the truncation error
and continue.  Some MPI error tests expect truncation errors,
but to continue after that.  Since we close the connection on
a truncation error, message data queued on the socket is lost.

This change copies the truncated message data into the app's buffer.
Previously, the receive was reported as a failure (ETRUNC), but
without the message data copied.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>