prov/tcp: Handle truncation of active message
Fix handling of the case where the user posts a buffer that's
too small to receive a message while the message is being
actively received into a saved buffer.

xnet_recv_saved() handles the case where the user posts a
receive buffer for a message that's currently being saved into
a provider buffer.  This is for unexpected message handling
where the receive was posted before we could fully read the
data from the tcp stream.

The code checks for message truncation and reports an error to
the user.  However, it doesn't reset the processing, so that the
excess message data is read from the socket.  This results in the
excess data being treated as a new message now that truncation
errors are no longer fatal to the connection.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>