prov/tcp: Fix spinning on fi_trywait()
When a completion is written to the CQ, the signaling fd is also written
to.  This wakes up any thread blocked on the cq.  However, that signal is
not cleared if the app is using fi_trywait() followed by blocking on the
cq's completion fd.  This can result in the application thread spinning
on fi_trywait() + fi_cq_read() loop.

Before returning from fi_trywait(), clear the signaling fd.  Also fail
the trywait call if there are any pending completions to read.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>