prov/sm2: Add RMA to SM2
RMA capabilties are added to SM2 using a new segmentation and reassembly
(SAR) protocol.  This allows RMA to read/write more than just the inject
size by using multiple SAR packets.

A SAR packet includes a context at the requestor to keep track of the
state of the transfer.  Currently the remote does not have any state and
relies on a LAST_PACKET flag to issue completions.

Although SAR can allow multiple packets in-flight at once, a corner case
of error-handling limits the count to 1 in many cases.  This restriction
may be removed in the future by adding a rendezvou and keeping state on
the remote as well.

Future changes will enable SAR protocol to be used for messages as well.

Signed-off-by: Luke Robison <lrbison@amazon.com>