prov/efa: rename DATA packet to CTSDATA packet
The packet type name DATA is too general and has caused
some confusion in the code. For example, the functions

	rxr_pkt_handle_data_XXX()

was meant to handle the DATA packet, not to handle
any packet has may contain data. Meanwhile, there
are function like efa_rdm_pke_handle_data_copied()
which is meant to handle any packet that has data.

This packet renamed the packet type to CTSDATA, because this
packet type is only used in long-CTS subprotocol.

The functions and data structures that are used for
CTSDATA packet have been renamed.

Signed-off-by: Wei Zhang <wzam@amazon.com>