prov/sm2: Introduce proto_flags
CMA host to device, SAR and RMA protocols in SM2 all need additional
flags to handle different messages internal to the protocol.

One approach is to use a different protocol for each message type but
this is cumbersome. Instead, we can set flags in the xfer entry that
indicate the message type.

To avoid alignment issues and keep the header size the same, we use
16 bits for the protocol and 16 bits for protocol flags.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>