intel/ci: improvement in build code for ucx.
changed the condition for building ucx.
Currently the code checked for "ucx" in
the path to build for ucx. This creates
issues if the branch name contains the string
"ucx". Also added it to provider list so that
logic can be reused and so that the
ucx flag is not misused.

Signed-off-by: Nikhil Nanal <nikhil.nanal@intel.com>