prov/sm2: Remove legacy tx_lock and switch to ep_lock
This patch removes the tx_lock and switches all locking to be using the
ep_lock.  This means that only one thread can be in a SM2 send/progress
operation at a time.

This change will synchronize the access to the freestack (which is
required for correctness).

This change will stop race conditions which can occur when thread 1
remaps in recv, while thread 2 is in the middle of a send.  Note: There
is still a race condition right now where thread 1 can call
fi_av_insert() and trigger a remap while thread 2 is in a send/recv and
cause an issue (will be fixed in later patch).

This change switches SM2 to use a gen_lock instead of a spin_lock which
will give a small performance improvement to the critical path.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>