prov/sm2: Remove progress from send path
We had a progress call in the send path to call progress if the send was
out of resources.  I did this so that an app that never calls progress
would still work.  The Libfabric API states that if an app returns
FI_EAGAIN, the application is responsible for calling progress.

This will also clean up some double locking issues that Coverity
reported.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>