prov/sm2: Allocate and map maximum size shm file at init
This change causes the entire shared memory file to be allocated and
mapped into each processes address space on startup. This change fixes
a multi-threaded race condition where one thread calls remap in
sm2_av_insert() while another thread is in the middle of a sm2_fifo_read(),
or sm2_fifo_write(). We need increased memory usage in order to provide
correctness to the user.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>