fabtests/common: separate RX and RMA counters
Previously, the same counter was used for receives and for
remote reads and writes. This caused an issue when using
counters with RMA tests since synchronization calls check
for completions up to rx_seq which are only incremented
when receives are posted. This allowed synchronization calls
to pass before they were ready.

Separating the counters, allows the synchronization calls to
properly compare the rx counter value with the rx_seq for
the posted receives only and not the RMA test operations.

The RMA trigger test is updated to remove unintuitive logic
to account for this discrepancy.

The RMA event test is update to explicitly call the counter
wait call with the rma counter instead of going through the
general completion call.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>