prov/util,efa,tcp,mrail,rxd,rxm,ucx: Refactor the usage of cntrs
Currently, cntrs for different operations have different
field names in util_ep, which causes code duplication and
makes it hard to expand in the future. This patch refactor
this code path by introducing two arrays: util_ep.cntrs, and
util_ep.cntr_inc_funcs to store the pointers to the cntrs
and its inc functions, which share the uniform procedure
in ep creation, bind, and close.

Signed-off-by: Shi Jin <sjina@amazon.com>