fabtests/benchmarks/bw: fix bw buffer utilization
This fixes a few issues in the benchmarking bandwidth test:
- Fix inject size check to be <= instead of <
- Fix buffer usage so that we send and receive into different
buffers, instead of always sending the same tx_buf and receiving
into the same rx_buf
- Fix extra fill_buf call - ft_inject also fills the tx buffer with
data so the test was doing it twice for all inject calls
- Add more thorough data verification. Using separate buffers allows
the test to check the contents of each separate transfer. Previously,
the rx_buf was only checked once which only verifies that the last
transfer had the correct contents.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>