prov/tcp: fix missing iov truncation on saved message path
Commit 489d7804cf78fe2cd6fe119aff595df5b8b8659c added truncation handling
in the saved message path but removed the iov truncation of the rx entry iov.
This caused erroneous truncation errors when a receive was posted that was
larger than the matched send because too much data would be read off the socket,
corrupting the tcp stream relative to the receiver.

This fixes a sporadic failure in the fi_rdm_tagged_bw test

Cherry-picked from commit 67cf6d97e49f570e77ac86ac148f893ca486a537

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>