fabtests/common: add inject function that posts a specific buffer
The regular send path has a post_tx_buf function that allows a test
to post a specific send buffer (not the default tx_buf). This patch
adds a comparable function for the inject path which currently only
has the ft_inject call which always posts the tx_buf as well as a
function similar to ft_post_tx which posts an inject with the default
buffer but without filling it with data.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>