include/osd.h: add defintion of memory ordering for no atomics
If no atomics are available, ofi will emulate atomics using locks.
In this case, no memory ordering is required.
If the atomics ordering primitives are not available, the build will fail.
This adds a definition for the memory ordering to allow it to build

Cherry-picked from commit 367b708b65cf89a9da74eade712cdd94402279ca

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>