prov/tcp: Add locking to trywait path for potential data race
Problems reported by thread sanitizer.  There is no locking
around checking the CQ or EQ for being empty.  It's unclear if
the locking is technically needed, but safer to add given the
code is not in a fast path (prior to the app making a blocking
call).

Note that with the change, the lock used by the CQ and cntr
signaling fd is not technically needed, as the progress
lock protects the entire operation.  Removal of the lock is
deferred for subsequent patches, as the EQ locking requires
further investigation (in order to remove).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>