prov/verbs: Track ep state to prevent duplicate shutdown events
Track the connection state of msg endpoints.  This is needed
to safely transition into a disconnect state, while ensuring that
we only report a shutdown event to the user once.

There are 2 separate ways that we can determine that an endpoint
has been disconnected.  One is by reading a disconnect event from
the rdma cm.  Another is by reading an error completion from the
CQ associated with the underlying QP.  Because we're not guaranteed
to get a disconnect event (the remote peer system may have crashed
without generating an RDMA disconnect traffic), if we read an
error completion from the CQ, we need to notify the user that the
local EP is no longer usable.  We do this by generating a local
shutdown event.

However, we _could_ read a disconnect event from the EQ.  We have
no way of knowing beforehand, or which order the disconnect
notification could occur (from a CQ or EQ event).  The state checks
allow us to transition the EP once and generate a single local
shutdown event.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>