contrib/intel/jenkins: Add a logging supported run_command
Adding run_logging_command()
This function allows CI processes to log their own output in the event that they aren't using slurm. This is necessary because current support grabs output from the invocation (runtests.py) command rather than the actual run_command. For tests like daos that don't use slurm, some of the output is missing. This command will allow the stdout of the child process to be piped into the log_file rather than the output of the runtests.py command.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>