fabtest: Fix reading addressing options
When running the following client/server apps:

fi_rdm_tagged_peek -p tcp -s 192.168.1.1
fi_rdm_tagged_peek -p tcp 192.168.1.1

The client app will segfault due to trying to insert a
NULL address into the AV.  The problem can be traced back
to ft_read_addr_opts() and the following if statement:

if (opts->dst_addr && (opts->src_addr || !opts->oob_port))

Whatever the intent of this check is isn't clear.  However,
the result is that because the client has only specified
a server address (opts->dst_addr is set, but opts->src_addr
is NULL), it gets skipped.  The code falls through
to the else clause, winds it way through libfabric calls,
and eventually ends up with fi_getinfo() returning
fi_info::dst_addr = NULL.  This address is then inserted
into the AV, which leads to the crash.

Fix by removing the odd extra requirement that we either need
to specify a src address with the dst address, or the oob_port
must be set.  A src address can either be set or not; the
code handles both cases.  And the use of oob sockets for
setup is independent of whether we're acting as a client or
server for the test.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>