prov/shm: Apply SAR check to only SAR protocol
The SAR status check happens in multiple places, this change makes it
only happen once. The SAR status check does not require additional
locking, this PR removes the additional locking that was only around one
instance of the SAR status check.

Reorganize smr_generic_rma() and smr_generic_sendmsg() so that we check
the SAR status before getting an entry on the command queue. This is
done so that the SAR protocol doesn't need to get the next entry, and
immediately return it every time the SAR status is not ready.

This SAR check is still happening in inject path, but cannot be moved
into the smr_do_sar() without fetching/returning a cmd queue entry (and
potentially hurting SAR performance).  We cannot move fetching the cmd
queue entry into the smr_do_sar() function because RMA functions re-use
the smr_do_sar() function and require access to the cmd entry before
do_sar function is called.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>