prov/shm: Replace tx_lock with ep_lock
The tx_lock allows a sender thread to send at the same time a receiver
thread can receive on an endpoint. This change gets rid of that
functionality and only allows 1 thread in an endpoint send or receive at
a time. When we move all SHM resources to the sender, we suspect that
the sender/receiver will be to intertwined to be able to progress
independently.

This change also boosts performance by switching to a ofi_genlock
instead of an ofi_spinlock.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>