prov/verbs: Prevent duplicate FI_SHUTDOWN events
A shutdown event can be generated after processing an RDMA_
CM_DISCONNECT_EVENT and if an error is detected on the CQ
(due to a failed transfers).  Because both could happen, the
user could see 2 FI_SHUTDOWN events.  Track if we've already
generated a shutdown event, so that we can skip reporting
it a second time.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>