prov/shm: Make SAR protocol handle 0 byte transfer.
Currently, SAR protocol doesn't handle 0 byte transfer
because it fails to update the resp->status correctly in
this corner case.

This issue can be triggered when application do send or write
with 0 byte msg, and requests FI_DELIVERY_COMPLETE (DC) in op flags.
SHM will use SAR for transmissions with DC, and it won't generate
tx completion due to this bug.

This patch fixes this issue.

Signed-off-by: Shi Jin <sjina@amazon.com>