Try libnvidia-ml.so.1 if .so symlink missing, and skip NVML when not available
Local testing on Debian has shown a case where the libnvidia-ml-so symlink to
libnvidia-ml.so.1 doesn't exist (but libnvidia-ml.so.1 exists).  So, try to
dlopen the .so.1 library if the generic .so dlopen fails.  In either case, if
libnvidia-ml doesn't exist, but the CUDA shared libraries do, skip trying to
use the NVML routines for getting the device count and just proceed to using
the CUDA routine.

Signed-off-by: Quincey Koziol <qkoziol@amazon.com>
(cherry picked from commit 804be986ae9d8f5a7a0b81d1434e84347733a0c1)
Signed-off-by: Quincey Koziol <qkoziol@amazon.com>