prov/tcp: Remove unneeded and incorrect rx_entry init code
xnet_get_save_rx() initializes some of the rx_entry fields for
the received message.  However, the same fields are also set
later in xnet_start_recv().  Remove the duplicate initialization,
one of which is incorrect (setting the cq to the tx_cq, rather
than the rx_cq).  The incorrect setting ends up being harmless,
because the field is reset before being used.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>