prov/efa: Fix and improve test_efa_rdm_ep_getopt
Currently, in test_efa_rdm_ep_getopt, the opt_len isn't reset
after the fi_getopt call, which is a bug because the opt_len
could be set by fi_getopt as an output value and violates the
goal of the test.

Use num_opt_names to calculate the length of opt_names array
instead of hardcoding it as 9.

Break the test_efa_rdm_ep_getopt into two tests:
1. test_efa_rdm_ep_getopt_undersized_optlen
2. test_efa_rdm_ep_getopt_oversized_optlen

Makes it easier to see what failed at a glance when running unit tests.

Signed-off-by: Shi Jin <sjina@amazon.com>