prov/tcp: Negotiate support for rendezvous
Update the RDM CM protocol to support different protocol versions.

The current code does not check the version as part of CM handling.
Instead, the CM response simply echoes back the version carried in
the request.

To account for this, we set a bit in the version field when sending
a response message.  When that bit has been set, it indicates that
the peer properly negotiated the version.  For CM messages that are
simply echoed back, without the bit set, we know that we're dealing
with the current implementation (version 0), regardless of the
version value.

If the version is set to 1, we can enable the rendezvous protocol
between the peers.  This ensures wire compatibility.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>