prov/shm: Set threading domain attr to FI_THREAD_UNSPEC
SHM is using real mutex's by default instead of no-op locks.  This
change will cause SHM to use no-op locks, which will increase the
performance of the default case (non-multithreaded).  If a user wants to
run a multi-threaded application, they will need to set threading to be
FI_THREAD_SAFE.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>