prov/opx: Add Virtual Lane and Partition pkey (FI_OPX_SL and FI_OPX_PKEY)
Add OPX functionality for virtual lane and for partition keys.

FI_OPX_SL is a user provided ENV that will specify the service-level (and eventually
the Virtual Lane) of the HPC job

FI_OPX_PKEY is a user provided ENV that will specify the partition key of this HPC job.

Errors will print a warning with FI_WARN, but will not be considered fatal to the
HPC job.

Signed-off-by: Tim Thompson <tim.thompson@cornelisnetworks.com>