fabtests/multinode: Use FI_DELIVERY_COMPLETE for 'barrier'
Restructure the barrier code to require delivery complete.
Without that option, a rank could post a send, which is buffered
locally, but a cq entry is written.  On reading the completion,
the rank can proceed without trying to drive progress on the cq.
The result is that a peer rank can hang waiting for the data to
arrive.

We update barrier to execute through rank 0 only, rather than
using an all-to-all communication pattern.  This reduces the
messages for the barrier from O(n^2) to O(n), which is sufficient
for how the internal barrier is used (simple synchronization point).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>