prov/efa: Add EFA device version to handshake
The device version (0xEFA0, 0xEFA1, etc) can be used to deduce various
capabilities of the EFA device. This version identifier is expected to
be available via ibv_device_attr.vendor_part_id (or hw_ver) (See
ibv_query_device(3)).

The dedicated header struct for the device version aligns to the 8-byte
boundary, marking the remaining 4 bytes as reserved, as opposed to
simply padding. This is intentional in the event we decide to store
some other analogous data there from ibv_device_attr.

Signed-off-by: Darryl Abbate <drl@amazon.com>