prov/efa: Do not create SHM peer when SHM is disabled
When the user sets FI_EFA_ENABLE_SHM_TRANSFER=0, do not create a SHM
endpoint.  The current behavior is that we just don't send via SHM when
one sets FI_EFA_ENABLE_SHM_TRANSFER=0, but we still create and poll the
peer endpoint.  New behavior is to not create SHM endpoint when peer is
disabled.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>