core: Define optional_caps
Optional capabilities may optionally be requested by an application. If
requested, providers are not required to support the capability. If
providers do not support the capability, the capability will be cleared
in the corresponding fi_info caps fields returned from fi_getinfo. Else,
providers will set the capability in the corresponding fi_info caps.

All capabilities, both primary and secondary, are eligible as being
optional.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>