prov/util: Revert util_av, util_cq and util_cntr to mutex
Switch ep_list_lock to always be a mutex lock until we can agree on a
better long term decision. The issue is that some endpoints want the
ability to not require FI_THREAD_SAFE at the domain level, but they want
to be able to control the endpoint in a seperate thread.  The
ep_list_lock is used for both the data path, and the control path.  In
order to adhere to the current api, and not break providers that follow
the current api, we need to revert this lock to a mutex until we can get
a optimal solution for everyone.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>