fabtests/hmem_ze: allocate one cq and cl on init
The fabtests ZE code is very inefficient as it creates a command queue
on every device (even though fabtests only uses one at a time) and
it also creates and frees a command list for every single copy call even
though it can be reused every time.

This patch cleans up the code so we only have a single command queue and
command list that we initialize on startup (on the device specified in
opts.device) which we reset on every copy call so that we can reuse it
throught the test.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>