fabtests/common: use common device host buffer for check_buf
The tx_msg_buf is a host buffer allocated with the hmem interface
which is used as a staging buffer for validation data to copy into the
device transfer buffer once filled with data. Currently, the tx_msg_buf
is onsly used on the fill_buf path but it can also be used on check_buf,
eliminating the need to alloc and free the staging buffer on every
transfer.

This also renames the tx_msg_buf variable to a more generic name
(dev_host_buf) since it is used on both the send and receive paths

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>