contrib/intel: simplify and expand ZE testing
Since not all of the fabtests supported device memory, the CI used
a wrapper script to run a subset of tests that were supported.
Now that all of the runfabtests support FI_HMEM properly, this simplifies
the ZE testing by going through the regular run_fabtests path.
The --device parameter is removed and replaced with the --way parameter
which indicates which direction to test (h2d, d2d, xd2d, default None).
This simplifies the code path and enables the ZE testing to run the
full testsuite, rather than the subset, increasing our coverage.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>