prov/ucx: Add FI_HMEM support
Device memory access is handled by UCX internally. UCX by default
uses a memtype cache to speed up identification of where a buffer
belongs to. Memory allocator hooks are installed to keep the memtype
cache updated.

Not much is needed at the provider level except for adding FI_HMEM
to the supported caps. The remaining changes are mostly workarounds
for the cases that the memory allocator hooks may not always work
as expected and thus the memtype cache may not have the information
for the device memory allocation:

* When device memory is registered, check if the allocation is
  already in the memtype cache. Try updating the cache if the
  information is missing.

* Enable FI_MR_HMEM mode if the application supports that. This
  requires all device memory allocation go through the memory
  registratin process and thus ensure the memtype cache is updated
  accordingly.

Removed a few unused headers.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>