prov/shm: Only copy header + msg on unexpected path
Currently, we always copy the entire command message of 256 bytes on
the unexpected path. For inline, we only need the header + size.  For
inject, we only need the header.  SAR, IOV, and IPC use the data field
in the message, so we continue to copy the entire command message for
them (for large message sizes, optimization is less important).

Signed-off-by: Seth Zegelstein <szegel@amazon.com>