fabtests/memcopy-xe: Fix data verification error for device buffer
Previously MAP_PRIVATE was used to map the device buffer for CPU access.
It used to work but with the latest Intel GPU driver update the caller
would see its own COW pages instead of the device memory. As a result,
the data verification at the end of the test would fail.

Changing the map type to MAP_SHARED fixes the issue. It should also work
with older drivers.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>